Split core.treeModel.delta.transform into two files, one with API and one with transformation cases. Renamed transform#addSpecialCase to transform#addTransformationCase.
