+++ tests/treemodel/delta/transform/weakinsertdelta.js
@@ -7,7 +7,7 @@
 
 'use strict';
 
-import transform from '/ckeditor5/core/treemodel/delta/transform.js';
+import transform from '/ckeditor5/core/treemodel/delta/transform/transform.js';
 
 import Text from '/ckeditor5/core/treemodel/text.js';
 import Position from '/ckeditor5/core/treemodel/position.js';
